Use path-clean instead of PathBuf::canonicalize
This fixes issues where canonicalize() would return an error because the path
didn't exist, but we didn't actually want to check that, we just wanted to
clean things like "./" and "/.." out of the path.

This change also calls clean() in some additional places to ensure consistent
paths. We no longer prefix all paths with "./" but if we have a dir
representing just the current dir it will be ".".