Add a new run_once config item for commands
This commit also fixes a bug where an on_dir filter would incorrectly run when
the directory contained a file that matched both its include and exclude
rules (for example `include = "**/*.go"` and `exclude = "foo/bar/*.go"`. The
include rule would win over exclude, which is the opposite of how this should
work.