Set root to cwd if it contains config file, and use cwd as root for basepaths
If we run precious in a directory that contains a config file, we should just
use that, even if we're in a VCS repo. Similarly, when asked to tidy or lint
files (or --all), we should always do that relative to the current directory,
not the root.

The previous behavior can be incredibly confusing. For example, I was trying
to integrate precious with my dzil-based Perl distros. The dzil build dir is a
subdirectory of the VCS root, so precious would look for applicable files
relative to the VCS root, but then run commands like `perlcritic` relative to
the _current_ directory. So if a file was in the VCS root but no the build dir
it led to very confusing failures.