Change how some errors are shown to user
Previously all errors were displayed using the log crate's `error!` macro. But
that macro would add a bunch of formatting to the error that is helpful in a
logging context, but not useful for users, like the module where the error
occurred and the error level.

Instead, _expected_ errors are now handled separately. Specifically, we do the
following:

If a linter indicates that a file failed a check, we still show this with the
poo emoji, along with any stdout or stderr from the linter inline with the
rest of the precious output.

If a linter or tidier fails for some other reason, either because we cannot
find the command to run in the PATH, or because running that command returns
an expected error _or_ prints to stderr unexpectedly, we show these errors in
the summary at the _end_ of the run.

In addition, the end of run errors are printed out without using `error!`, so
the coloring and display format is entirely controlled at the printing site.

This should finish addressing #10.