Do a major refactoring of the precious.rs code
I realized that the pattern I had in this code of calling a method like
`self.root()` and having that internally calculate and set `self.root` wasn't
very Rust-y. It's a pattern I borrowed from Moose's (in Perl) lazy builders.

But it's a lot simpler to set all the relevant Precious struct attributes up
front in the constructor.

This also had the side benefit of removing a _lot_ of fallibility from deep in
the call stack, so I could simplify quite a few return types and remove a lot
of `?`.

This also meant changing basepaths a bit so that CLI paths were passed into
its `paths()` instead of being stored in the `BasePaths` struct.