Fix the testhelper code to work in the presence of a totally clean git config
If the testhelper tries to `git commit` in an environment where git hasn't
been configured at all, you get an error like this:

```
*** Please tell me who you are.

Run

  git config --global user.email "you@example.com"
  git config --global user.name "Your Name"

to set your account's default identity.
Omit --global to set the identity only in this repository.
```

So we need to set these config values for the test repo.

Fixes #15.