Major refactor in support of running precious in a subdir
The refactor is so I can add integration tests that depend on the command and
testhelper modules without making the testhelper module available outside of a
`#cfg[test]` directive.

This splits the code into several crates, one of which is the binary and one
of which is a set of integration tests. These tests cover running precious in
a subdir.