More path handling cleanups
We now call `fs::canonicalize` on all paths in addition `.clean`. I think this
will fix the issues on Windows.

I also renamed a number of methods and vars to make it clearer what they
are.

Finally, I changed the `files_from_cli` method so it doesn't relativize the
files twice.