Get integration tests passing on Windows
This required some more path fiddling.

Also, the file locking in Perl didn't seem to work on Windows. So now instead
of locking a single output file, we have `some-linter` write one file per
invocation, each with a unique name containing the PID of `some-linter`. Then
we simply concatenate all the files' content together and work with that.