Major refactoring of internals
Writing output is now done in its own thread. This is done by sending messages
to that thread down a crossbeam channel. This allows for different output
formats, including formats that need to first buffer events before printing
output (like XML).

This also hides some of the `Precious` struct internals so that it can only be
created from an `App` instance.

Finally, the `Exit` type is now an enum in its own module.