Make `perlimports` precede `perltidy` in generated config for Perl
The `perlimports` command might change the formatting of a page in a way that
`perltidy` might later want to reformat. Running `perlimports` after
`perltidy` meant that tidying Perl code could leave the code in a state where
it fails linting checks.

This PR changes the generated config so that `perlimports` is run first.