Clean up some issues with generated Go config
This replaces the use of `golangci-lint` for formatting with `gofumpt`. It's
not really viable to use `golangci-lint` as the primary formatter.

However, we still configure `golangci-lint` as a tidier with its `--fix` flag,
since many of the linters it bundles have the ability to fix the issues they find.

This also changes the `golangci-lint` config file to `.golangci.yml` instead
of `golangci-lint.yml`.